backend/python/requirements.txt
alembic==1.13.1
appdirs==1.4.4
attrs==21.2.0
black==20.8b1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython>=2.0.0
firebase-admin==4.5.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.22
asyncpg==0.29.0
email-validator==2.1.0
# jinja2==3.1.2 - removed due to conflict with Jinja2==2.11.2
python-multipart==0.0.6
pydantic-settings>=2.0.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
mongoengine==0.23.0
mongomock==3.23.0
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.4
pytest-mock==3.6.1
python-dateutil==2.8.1
python-dotenv>=0.21.0
python-editor==1.0.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
sentinels==1.0.0
six==1.15.0
SQLAlchemy==2.0.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions>=4.0.0
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
